added GPS lat lon to gather_gps()
